requirements.txt
Flask==2.0.2
google-cloud==0.34.0
google-cloud-dlp==3.6.0
google-cloud-speech==2.12.0
google-cloud-storage==2.1.0
protobuf==3.19.4
pydantic==1.9.0
python-dotenv==0.19.2
requests==2.27.1
